refactor to now be useable to execute. Trajectory evaluation is different (it should not be)
